VACMS-16348: Moves work out of backend and vamc and into facilities
